// Angular-RecipeBook-FrontEnd/src/app/core/error-handler/enums/api-exception-code.enum.ts
export const enum ApiExceptionCode {
  UserNotFound = 1,
  SignInUserNotFound = 2,
  UnauthorizedUser = 3,
  InvalidAuthorizedUserId = 4,
  UserIsLockedOut = 5,
  UserIsNotAllowed = 6,
  LoginFailed = 7,
  RecipeNotFound = 8,
  UpgradeableRecipeNotFound = 9,
  DeletableRecipeNotFound = 10,
  RecipeNameIsAlreadyExist = 11,
}